requirements.txt
numpy==1.23.1
nuscenes-devkit==1.1.10
opencv-python==4.6.0.66
Pillow==9.2.0